restructure pydantic and sqlalchemy models for apis
