+++ backend/api/personAPI/sqlalchemy_models.py
@@ -1,0 +1,33 @@
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import Column, Integer, String, Boolean, Date, ForeignKey
+
+Base = declarative_base()
+
+class Person(Base):
+    __tablename__ = "Person"
+    PersonID = Column(Integer, primary_key=True)
+    Name = Column(String(50))
+    Occupation = Column(String(25))
+    DoB = Column(Date)
+    Nationality = Column(String(25))
+    Description = Column(String(255))
+    CountryOfResidency = Column(String(50))
+    PEPStatus = Column(Boolean)
+    SourceOfWealth = Column(String(25))
+    ImgURL = Column(String(255))
+
+
+class Company(Base):
+    __tablename__ = "Company"
+    CompanyID = Column(Integer, primary_key=True)
+    Name = Column(String(100))
+    Industry = Column(String(25))
+    CountryOfHeadquarters = Column(String(50))
+    Type = Column(String(25))
+
+
+class PersonCompany(Base):
+    __tablename__ = "Person-Company"
+    PersonID = Column(Integer, ForeignKey("Person.PersonID"), primary_key=True)
+    CompanyID = Column(Integer, ForeignKey("Company.CompanyID"), primary_key=True)
+    Role = Column(String(25))
