+++ src/app/home/page.js
@@ -2,6 +2,7 @@
 import SearchBar from "@/components/SearchBar";
 import { useState } from "react";
 import { useRouter } from "next/navigation";
+import Image from "next/image";
 
 export default function Home() {
   const [search, setSearch] = useState("");
@@ -19,7 +20,7 @@ export default function Home() {
   return (
     <div className="flex h-screen justify-center items-center">
       <div className="w-full flex-col flex justify-center items-center">
-        <img src="/logo.png" className="sm:w-1/3 w-1/2 m-3"></img>
+        <Image width={1000} height={1000} className="sm:w-1/3 w-1/2 m-3" src="/logo.png"></Image>
         <SearchBar
           searchQuery={search}
           onSearchChange={handleSearchChange}
